Refactor thread message components to support TextMessage and JsonMessage
